NTFS: improve attribute debug logging.
